Small engine api client creation refactoring
