+++ app/src/main/kotlin/maru/app/Helpers.kt
@@ -21,8 +21,12 @@ import kotlin.io.path.Path
 import kotlin.time.Duration.Companion.minutes
 import kotlin.time.toJavaDuration
 import maru.config.ApiEndpointConfig
+import maru.consensus.ElFork
+import maru.executionlayer.client.ExecutionLayerEngineApiClient
+import maru.executionlayer.client.PragueWeb3JJsonRpcExecutionLayerEngineApiClient
 import tech.pegasys.teku.ethereum.executionclient.auth.JwtConfig
 import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JClient
+import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JExecutionEngineClient
 import tech.pegasys.teku.ethereum.executionclient.web3j.Web3jClientBuilder
 import tech.pegasys.teku.infrastructure.time.SystemTimeProvider
 
@@ -45,4 +49,15 @@ object Helpers {
       .timeProvider(SystemTimeProvider.SYSTEM_TIME_PROVIDER)
       .executionClientEventsPublisher {}
       .build()
+
+  fun buildExecutionEngineClient(
+    endpoint: ApiEndpointConfig,
+    elFork: ElFork,
+  ): ExecutionLayerEngineApiClient {
+    val web3JEngineApiClient: Web3JClient = createWeb3jClient(endpoint)
+    val web3jExecutionLayerClient = Web3JExecutionEngineClient(web3JEngineApiClient)
+    return when (elFork) {
+      ElFork.Prague -> PragueWeb3JJsonRpcExecutionLayerEngineApiClient(web3jExecutionLayerClient)
+    }
+  }
 }
