+++ app/src/main/kotlin/maru/app/QbftProtocolFactoryWithBeaconChainInitialization.kt
@@ -16,24 +16,14 @@
 package maru.app
 
 import java.time.Clock
-import maru.config.ApiEndpointConfig
 import maru.config.MaruConfig
-import maru.consensus.ElFork
 import maru.consensus.ForkSpec
 import maru.consensus.NewBlockHandler
 import maru.consensus.NextBlockTimestampProvider
 import maru.consensus.ProtocolFactory
-import maru.consensus.StaticValidatorProvider
-import maru.consensus.blockimport.BlockBuildingBeaconBlockImporter
-import maru.consensus.blockimport.SealedBeaconBlockImporter
-import maru.consensus.blockimport.TransactionalSealedBeaconBlockImporter
-import maru.consensus.qbft.ProposerSelector
-import maru.consensus.qbft.ProposerSelectorImpl
 import maru.consensus.qbft.QbftConsensusConfig
 import maru.consensus.qbft.QbftProtocolFactory
 import maru.consensus.state.FinalizationState
-import maru.consensus.state.StateTransition
-import maru.consensus.state.StateTransitionImpl
 import maru.core.BeaconBlock
 import maru.core.BeaconBlockBody
 import maru.core.BeaconBlockHeader
@@ -44,19 +34,13 @@ import maru.core.SealedBeaconBlock
 import maru.core.Validator
 import maru.crypto.Crypto
 import maru.database.BeaconChain
-import maru.executionlayer.client.ExecutionLayerEngineApiClient
-import maru.executionlayer.client.PragueWeb3JJsonRpcExecutionLayerEngineApiClient
-import maru.executionlayer.manager.ExecutionLayerManager
 import maru.executionlayer.manager.JsonRpcExecutionLayerManager
 import maru.mappers.Mappers.toDomain
 import maru.serialization.rlp.RLPSerializers
 import maru.serialization.rlp.stateRoot
-import org.hyperledger.besu.consensus.common.bft.ConsensusRoundIdentifier
 import org.hyperledger.besu.plugin.services.MetricsSystem
 import org.web3j.protocol.Web3j
 import org.web3j.protocol.core.DefaultBlockParameter
-import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JClient
-import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JExecutionEngineClient
 
 class QbftProtocolFactoryWithBeaconChainInitialization(
   private val maruConfig: MaruConfig,
@@ -110,35 +94,6 @@ class QbftProtocolFactoryWithBeaconChainInitialization(
     updater.commit()
   }
 
-  private fun createSealedBeaconBlockImporter(
-    executionLayerManager: ExecutionLayerManager,
-    proposerSelector: ProposerSelector,
-    localNodeIdentity: Validator,
-    beaconChain: BeaconChain,
-    stateTransition: StateTransition,
-  ): SealedBeaconBlockImporter {
-    val finalizationStateProvider = { beaconBlockBody: BeaconBlockBody ->
-      val hash = beaconBlockBody.executionPayload.blockHash
-      FinalizationState(hash, hash)
-    }
-    val shouldBuildNextBlock =
-      { beaconState: BeaconState, roundIdentifier: ConsensusRoundIdentifier ->
-        val nextProposerAddress =
-          proposerSelector.getProposerForBlock(beaconState, roundIdentifier).get().address
-        nextProposerAddress.contentEquals(localNodeIdentity.address)
-      }
-    val blockBuilderIdentity = Crypto.privateKeyToValidator(maruConfig.validator!!.privateKey)
-    val beaconBlockImporter =
-      BlockBuildingBeaconBlockImporter(
-        executionLayerManager = executionLayerManager,
-        finalizationStateProvider = finalizationStateProvider,
-        nextBlockTimestampProvider = nextTargetBlockTimestampProvider,
-        shouldBuildNextBlock = shouldBuildNextBlock,
-        blockBuilderIdentity = blockBuilderIdentity,
-      )
-    return TransactionalSealedBeaconBlockImporter(beaconChain, stateTransition, beaconBlockImporter)
-  }
-
   override fun create(forkSpec: ForkSpec): Protocol {
     require(forkSpec.configuration is QbftConsensusConfig) {
       "Unexpected fork specification! ${
@@ -149,7 +104,7 @@ class QbftProtocolFactoryWithBeaconChainInitialization(
     val qbftConsensusConfig = forkSpec.configuration as QbftConsensusConfig
 
     val engineApiExecutionLayerClient =
-      buildExecutionEngineClient(
+      Helpers.buildExecutionEngineClient(
         maruConfig.validator!!.engineApiClient,
         qbftConsensusConfig.elFork,
       )
@@ -162,18 +117,6 @@ class QbftProtocolFactoryWithBeaconChainInitialization(
       initializeDb(beaconChain.newUpdater())
     }
 
-    val localNodeIdentity = Crypto.privateKeyToValidator(maruConfig.validator!!.privateKey)
-    val validatorProvider = StaticValidatorProvider(setOf(localNodeIdentity))
-    val stateTransition = StateTransitionImpl(validatorProvider)
-    val proposerSelector = ProposerSelectorImpl
-    val sealedBeaconBlockImporter =
-      createSealedBeaconBlockImporter(
-        executionLayerManager = executionLayerManager,
-        beaconChain = beaconChain,
-        proposerSelector = proposerSelector,
-        localNodeIdentity = localNodeIdentity,
-        stateTransition = stateTransition,
-      )
     val qbftProtocolFactory =
       QbftProtocolFactory(
         beaconChain = beaconChain,
@@ -182,24 +125,9 @@ class QbftProtocolFactoryWithBeaconChainInitialization(
         finalizationStateProvider = finalizationStateProvider,
         nextBlockTimestampProvider = nextTargetBlockTimestampProvider,
         newBlockHandler = newBlockHandler,
-        sealedBeaconBlockImporter = sealedBeaconBlockImporter,
         executionLayerManager = executionLayerManager,
-        stateTransition = stateTransition,
-        proposerSelector = proposerSelector,
-        validatorProvider = validatorProvider,
         clock = clock,
       )
     return qbftProtocolFactory.create(forkSpec)
-  }
-
-  private fun buildExecutionEngineClient(
-    endpoint: ApiEndpointConfig,
-    elFork: ElFork,
-  ): ExecutionLayerEngineApiClient {
-    val web3JEngineApiClient: Web3JClient = Helpers.createWeb3jClient(endpoint)
-    val web3jExecutionLayerClient = Web3JExecutionEngineClient(web3JEngineApiClient)
-    return when (elFork) {
-      ElFork.Prague -> PragueWeb3JJsonRpcExecutionLayerEngineApiClient(web3jExecutionLayerClient)
-    }
   }
 }
