+++ e2e/src/acceptanceTest/kotlin/maru/e2e/CliqueToPosTest.kt
@@ -30,6 +30,7 @@ import kotlin.time.toJavaDuration
 import maru.app.Helpers
 import maru.app.MaruApp
 import maru.config.ApiEndpointConfig
+import maru.consensus.ElFork
 import maru.consensus.NewBlockHandler
 import maru.consensus.blockimport.FollowerBeaconBlockImporter
 import maru.core.BeaconBlock
@@ -211,7 +212,7 @@ class CliqueToPosTest {
   }
 
   private fun buildBlockImportHandler(engineApiConfig: ApiEndpointConfig): NewBlockHandler<ForkChoiceUpdatedResult> =
-    FollowerBeaconBlockImporter.create(Helpers.createWeb3jClient(engineApiConfig))
+    FollowerBeaconBlockImporter.create(Helpers.buildExecutionEngineClient(engineApiConfig, ElFork.Prague))
 
   private fun waitTillTimestamp(timestamp: Long) {
     await.timeout(1.minutes.toJavaDuration()).pollInterval(5.seconds.toJavaDuration()).untilAsserted {
